// app/src/app/bo/models/feature.ts
export class Feature {
  id?: string | undefined;
  mainTextA?: string | undefined;
  descriptionA?: string | undefined;
  buttonTextA?: string | undefined;
  buttonLinkA?: string | undefined;
  mainTextB?: string | undefined;
  descriptionB?: string | undefined;
  buttonTextB?: string | undefined;
  buttonLinkB?: string | undefined;
  mainTextC?: string | undefined;
  descriptionC?: string | undefined;
  buttonTextC?: string | undefined;
  buttonLinkC?: string | undefined;
}